run epatch_user (if possible) for every package
This makes testing patches easier and allows the use of local patchsets
for various packages.

Depending on the eclasses included, epatch_user is run by the default
prepare function; but most packages that override prepare do not
manually run epatch_user. This commit causes it to be run before
whatever prepare function the package includes. Portage uses a flag
file to ensure that epatch_user is not run multiple times, so this is
safe even for ebuilds that manually run it.

Note that if the ebuild does not inherit the eclass that contains
epatch_user, the function will not be available, and will not be run.