add configuration for qemu
qemu has a USE_EXPAND variable for which targets to enable. Since I'm
only using it for servers (not e.g., development) I will only be using
KVM on the native architecture.

fdt is not useful for x86.
tls is for VNC security.
usb and usbredir are required for virt-manager.
virtfs is so I can use BTRFS features and avoid COW problems in VMs.