pulseaudio: Disable OSS support
It is useless anyway, since I have no legacy OSS-only applications. The
main reason for disabling it, though, is that it is completely broken on
musl because it overrides functions from the libc. It would be more
correct to modify the ebuild, but using epatch_user lets me avoid
maintaining an overlay ebuild.