Add netbeans; factor out java support
I need netbeans 8.1 with java and javafx for one class,
and C support for another class. This is all of the fun
stuff required to get that working. Since icedtea 1.8 is
not released, to get java 8 (which I also must have), I
have to use the Oracle JDK.

dev-java/guava has a hard dependency on a Java 7 JDK. It
requires this patch to build on Java 8, and modifying the
ebuild to change the dependencies (--nodeps does not work
because of the auto-dependency stuff in the java-utils-2
eclass).