mask default virtual providers since portage is stupid
It can't seem to figure out, that if I tell it to only use binary
packages, that it should install the dependency that *can* be satisfied
with a binary package, instead of erroring out because the first
alternative has no binary package available.

I'm not actually interested in masking these, but emerge -K doesn't
figure out the existence of vim and nail otherwise.